Fix errors from Clang Tidy
error: pointer parameter 'setupFail' can be pointer to const
error: narrowing conversion from 'unsigned long' to signed type 'int' is implementation-defined
error: narrowing conversion from 'double' to 'int'
error: narrowing conversion from 'size_t' (aka 'unsigned long') to signed type 'int' is implementation-defined
error: Value stored to 'pOUTIndex' is never read
error: Value stored to 'startSim' during its initialization is never read
error: duplicate include
error: result of integer division used in a floating point context; possible loss of precision
error: 1st function call argument is an uninitialized value
error: either cast from 'int' to 'size_t' (aka 'unsigned long') is ineffective, or there is loss of precision before the conversion
error: performing an implicit widening conversion to type 'size_t' (aka 'unsigned long') of a multiplication performed in type 'int'
error: the value returned by this function should not be disregarded; neglecting it may lead to errors
error: Value stored to 'failEarly' is never read
error: redundant explicit casting to the same type 'int' as the sub-expression, remove this casting
error: Array access results in a null pointer dereference

+++ src/SW_Weather.c
@@ -785,7 +785,7 @@ void SW_WTH_setWeatherValues(
     TimeInt yearIndex;
     TimeInt doy;
     TimeInt tempDoy;
-    int tmp;
+    double tmp;
 
     Bool useMaxMinTemp = (Bool) (inputFlags[TEMP_MAX] && inputFlags[TEMP_MIN]);
     Bool hasMaxMinTempValues = swFALSE;
@@ -1125,7 +1125,7 @@ the weather history for every possible input variable
 */
 void allocate_temp_weather(
     TimeInt nYears,
-    int extraStorMult,
+    size_t extraStorMult,
     double ****fullWeathHist,
     LOG_INFO *LogInfo
 ) {
@@ -2013,7 +2013,7 @@ void checkAllWeather(
 @note Used by rSOILWAT2
 */
 void clear_hist_weather(
-    int extraStorMult, SW_WEATHER_HIST *yearWeather, double **fullWeathHist
+    size_t extraStorMult, SW_WEATHER_HIST *yearWeather, double **fullWeathHist
 ) {
     /* --------------------------------------------------- */
     TimeInt d;
